// src/app/models/Delivery.interface.ts
export interface Delivery {
  id: string;
  documento: string;
  motorista: { nome: string };
  cliente_origem: { nome: string, endereco: string, bairro: string, cidade: string };
  cliente_destino: { nome: string, endereco: string, bairro: string, cidade: string };
  status_entrega: string;
}